refactor(core): split model download into progress, variant, download
The former model/mod.rs god-file is now a directory module. Public
gigastt_core::model::* paths are unchanged. CI model-cache keys hash
the whole model/ directory so pin changes in variant.rs still bust
the cache.
