+++ .github/workflows/ci.yml
@@ -241,7 +241,7 @@ jobs:
         uses: actions/cache@v6
         with:
           path: ~/.gigastt/models
-          key: gigastt-model-v3-onnx-${{ hashFiles('crates/gigastt-core/src/model/mod.rs') }}
+          key: gigastt-model-v3-onnx-${{ hashFiles('crates/gigastt-core/src/model/**') }}
 
       - name: Download model
         if: github.ref == 'refs/heads/main' && github.event_name == 'push' && steps.cache-model.outputs.cache-hit != 'true'
@@ -352,7 +352,7 @@ jobs:
         uses: actions/cache@v6
         with:
           path: ~/.gigastt/models
-          key: gigastt-model-v3-onnx-${{ hashFiles('crates/gigastt-core/src/model/mod.rs') }}
+          key: gigastt-model-v3-onnx-${{ hashFiles('crates/gigastt-core/src/model/**') }}
 
       - name: Download model
         if: steps.cache-model.outputs.cache-hit != 'true'
@@ -384,7 +384,7 @@ jobs:
         uses: actions/cache@v6
         with:
           path: ~/.gigastt/models
-          key: gigastt-model-v3-onnx-${{ hashFiles('crates/gigastt-core/src/model/mod.rs') }}
+          key: gigastt-model-v3-onnx-${{ hashFiles('crates/gigastt-core/src/model/**') }}
 
       - name: Download model
         if: steps.cache-model.outputs.cache-hit != 'true'
@@ -869,7 +869,7 @@ jobs:
         uses: actions/cache@v6
         with:
           path: ~/.gigastt/models
-          key: gigastt-model-v3-onnx-${{ hashFiles('crates/gigastt-core/src/model/mod.rs') }}
+          key: gigastt-model-v3-onnx-${{ hashFiles('crates/gigastt-core/src/model/**') }}
 
       - name: Download model
         if: steps.cache-model.outputs.cache-hit != 'true'
